third_party/blink/web_tests/external/wpt/css/css-grid/grid-definition/grid-limits-001-expected.txt

This is a testharness.js-based test.
FAIL repeat(9999) should be supported in both axises assert_equals: repeat(9999, 1px) should make grid height 9999px expected "9999px" but got "1000px"
FAIL Implicit grid should support lines [-10k, 10k]. assert_equals: Positioning children at lines -10k and 10k should generate 19998 implicit 1px tracks. expected "19998px" but got "1000px"
Harness: the test ran to completion.
